Clean up seeding folder and update event seeding script
