--- database/seed/filtered_out_event_ids.csv
@@ -1,7 +1,0 @@
-id
-91
-129
-130
-133
-134
-375
